Lesson 6: 'Profile' class added, and implemented to 'profile' command
